pyproject.toml
[project]
name = "u2net-removal-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "uvicorn>=0.35.0",
]
